Add Rev::Listener#fileno to access the numeric descriptor
It is useful to access the underlying file descriptor in case a
server needs to exec() a new process and share the file
descriptor information with its children (via the command-line
or environment variable).   The child could then call
*Server.for_fd on the file descriptor.

This also makes the existing unix_listener_spec easier
to implement.

Signed-off-by: Eric Wong <normalperson@yhbt.net>