fix on_timer event_callback for TimerWatcher
There was a timer_watcher event_callback defined for the
IOWatcher class in the timer_watcher.rb file.  I assume this is
a copy-and-paste error and a grep of the source shows no
"timer_watcher" methods ever being dispatched.  So change this
to allow the "on_timer" method to be overidden when given a
block.

Also added a spec to show it works.