Detect fallback local loopback address correctly
In the previous PR, #71, local loopback address is
detected even though hosts file is empty, but
there is a case that hosts file is NOT empty
and localhost entry is missing. In such a case,
Coolio::DNSResolver.hosts("localhost") does not work
as expected. This PR should fix such case.